requirements.txt
fastapi
uvicorn
httpx
sqlparse
sqlglot
sqlalchemy
pymysql
pandas
plotly
kaleido
openpyxl
reportlab
tqdm
requests
streamlit
aiogram>=3.0
apscheduler>=3.10
python-dotenv>=1.0
